+++ index.js
@@ -46,7 +46,7 @@ export const handler = async (event) => {
     // Note: Some of these properties might not be relevant for certain cases, you can remove them if they are not relevant
     //       For example, metadata does not exist in interceptors' events.
     const { data, eventToken, eventTraceparent, metadata } = extractDataFromEvent(event);
-
+    
     // We extract the traceparent from the event
     // If the traceparent is not included, we create a new one
     traceparent = eventTraceparent ?? createTraceparent();