+++ src/interceptorPostEntity/perform.js
@@ -1,4 +1,4 @@
-import { callToAPIExample } from "../utils/callToAPIExample.js";
+import { searchOnlyFemalePatients } from "../services/biot/patient.js";
 
 export const perform = async (data, token, traceparent) => {
   // -----------------------------------------------------------------------------------------
@@ -19,7 +19,7 @@ export const perform = async (data, token, traceparent) => {
   // This is an example of calling a BioT API (using the token from the service users token)
   // In this case we are making a get patients request to organization API
 
-  const callExampleResponse = await callToAPIExample(token, traceparent);
+  const callExampleResponse = await searchOnlyFemalePatients(token, traceparent);
 
   // In this example you perform your logic with the response Here