+++ src/nonspecific/index.js
@@ -1,4 +1,4 @@
-import { authenticate, login } from "../utils/index.js";
+import { login, authenticate } from "../services/auth/index.js";
 import { extractDataFromEvent } from "./extractDataFromEvent.js";
 import { createErrorResponse } from "./createErrorResponse.js";
 import { perform } from "./perform.js";